Removed all unnecessary functions, bug fixes, fixing a master destroy node issue and some issues with the protocol queue, plus lock confusion. move load and run to ot2_client
